[Refactor] Review 관련 DTO - Model 분리
